Add Pokemon catching functionality to the Pokedex
This commit includes the addition of a new 'catch' command which allows users to catch a Pokemon by typing 'catch' followed by the Pokemon's name. Also added a 'caughtPokemon' map to store the caught pokemon in.

The new commandCatch function makes a GET request to the PokeAPI for a Pokemon and determines whether the user catches the Pokemon based on a random number. The caught pokemon is then stored in the caughtPokemon map.

Also included is a new 'GetPokemon' function in the PokeAPI package to fetch Pokemon data from the API.

The changes make the Pokedex tool more realistic and interactive by implementing a fundamental Pokemon feature which is the ability to catch pokemon.